Refactor io/ops into submodules (#254)
