+++ tests/unit/test_dask_nvt.py
@@ -26,7 +26,7 @@ from dask.dataframe import assert_eq
 
 from nvtabular import Dataset, Workflow
 from nvtabular import ops as ops
-from nvtabular.io import Shuffle
+from nvtabular.io.shuffle import Shuffle
 from tests.conftest import allcols_csv, mycols_csv, mycols_pq
