+++ tests/unit/test_io.py
@@ -28,7 +28,7 @@ from dask.dataframe import assert_eq
 import nvtabular as nvt
 import nvtabular.io
 from nvtabular import ops as ops
-from nvtabular.io import ParquetWriter
+from nvtabular.io.parquet import ParquetWriter
 from tests.conftest import allcols_csv, mycols_csv, mycols_pq
